requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
google-auth==2.23.2
google-auth-oauthlib==1.1.0
gspread==5.11.3
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
iniconfig==2.0.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.1
phonenumberslite==8.13.22
pluggy==1.3.0
psycopg2-binary==2.9.8
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.5
